+++ src/NeuralNet/Optimizers/Momentum.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\NeuralNet\Optimizers;
 
 use Tensor\Tensor;
+use Rubix\ML\Helpers\Params;
 use Rubix\ML\NeuralNet\Parameter;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 
@@ -11,13 +12,12 @@ use function get_class;
 /**
  * Momentum
  *
- * Momentum adds velocity to each step until exhausted. It does so by accumulating
- * momentum from past updates and adding a factor of the previous velocity to the
- * current step.
+ * Momentum adds velocity to each step until exhausted. It does so by accumulating momentum from past updates and adding
+ * a factor of the previous velocity to the current step.
  *
  * References:
- * [1] D. E. Rumelhart et al. (1988). Learning representations by back-propagating
- * errors.
+ * [1] D. E. Rumelhart et al. (1988). Learning representations by back-propagating errors.
+ * [2] I. Sutskever et al. (2013). On the importance of initialization and momentum in deep learning.
  *
  * @category    Machine Learning
  * @package     Rubix/ML
@@ -40,6 +40,13 @@ class Momentum implements Optimizer, Adaptive
     protected float $decay;
 
     /**
+     * Should we employ Nesterov's lookahead (NAG) when updating the parameters?
+     *
+     * @var bool
+     */
+    protected bool $lookahead;
+
+    /**
      * The parameter cache of velocity matrices.
      *
      * @var \Tensor\Tensor[]
@@ -51,9 +58,10 @@ class Momentum implements Optimizer, Adaptive
     /**
      * @param float $rate
      * @param float $decay
+     * @param bool $lookahead
      * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
-    public function __construct(float $rate = 0.001, float $decay = 0.1)
+    public function __construct(float $rate = 0.001, float $decay = 0.1, bool $lookahead = false)
     {
         if ($rate <= 0.0) {
             throw new InvalidArgumentException('Learning rate must be'
@@ -67,6 +75,7 @@ class Momentum implements Optimizer, Adaptive
 
         $this->rate = $rate;
         $this->decay = $decay;
+        $this->lookahead = $lookahead;
     }
 
     /**
@@ -101,6 +110,11 @@ class Momentum implements Optimizer, Adaptive
 
         $this->cache[$param->id()] = $velocity;
 
+        if ($this->lookahead) {
+            $velocity = $gradient->multiply($this->rate)
+                ->add($velocity->multiply(1.0 - $this->decay));
+        }
+
         return $velocity;
     }
 
@@ -111,6 +125,7 @@ class Momentum implements Optimizer, Adaptive
      */
     public function __toString() : string
     {
-        return "Momentum (rate: {$this->rate}, decay: {$this->decay})";
+        return "Momentum (rate: {$this->rate}, decay: {$this->decay},"
+            . ' lookahead: ' . Params:: toString($this->lookahead) . ')';
     }
 }