+++ src/Transformers/LinearDiscriminantAnalysis.php
@@ -9,6 +9,7 @@ use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Traits\AutotrackRevisions;
 use Rubix\ML\Specifications\ExtensionIsLoaded;
+use Rubix\ML\Specifications\SpecificationChain;
 use Rubix\ML\Specifications\ExtensionMinimumVersion;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
@@ -63,9 +64,10 @@ class LinearDiscriminantAnalysis implements Transformer, Stateful, Persistable
      */
     public function __construct(int $dimensions)
     {
-        if (ExtensionIsLoaded::with('tensor')->passes()) {
-            ExtensionMinimumVersion::with('tensor', '2.1.4')->check();
-        }
+        SpecificationChain::with([
+            new ExtensionIsLoaded('tensor'),
+            new ExtensionMinimumVersion('tensor', '2.1.4'),
+        ])->check();
 
         if ($dimensions < 1) {
             throw new InvalidArgumentException('Dimensions must be'