+++ src/Transformers/PrincipalComponentAnalysis.php
@@ -8,6 +8,7 @@ use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Traits\AutotrackRevisions;
 use Rubix\ML\Specifications\ExtensionIsLoaded;
+use Rubix\ML\Specifications\SpecificationChain;
 use Rubix\ML\Specifications\ExtensionMinimumVersion;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
 use Rubix\ML\Exceptions\InvalidArgumentException;
@@ -74,9 +75,10 @@ class PrincipalComponentAnalysis implements Transformer, Stateful, Persistable
      */
     public function __construct(int $dimensions)
     {
-        if (ExtensionIsLoaded::with('tensor')->passes()) {
-            ExtensionMinimumVersion::with('tensor', '2.1.4')->check();
-        }
+        SpecificationChain::with([
+            new ExtensionIsLoaded('tensor'),
+            new ExtensionMinimumVersion('tensor', '2.1.4'),
+        ])->check();
 
         if ($dimensions < 1) {
             throw new InvalidArgumentException('Dimensions must be'