+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+- 1.1.0
+    - Update to Scienide Tensor 3.0
+    - Added Nesterov's lookahead to Momentum Optimizer
+    - Added Reversible transformer interface
+    - MaxAbs, Z Score, and Robust scalers are now Reversible
+    - Min Max Normalizer now implements Reversible
+    - TF-IDF Transformer is now Reversible
+    - Added Preset cluster seeder
+    - Added Concatenator extractor
+
 1.0.3
     - Do not remove `groups` property from symbol table